Added new module queuehandling with functions that do queue manipulations. Moved existing relevant functions from auxiliary module into the new one. Added doc file as well.
