+++ pilot/util/auxiliary.py
@@ -156,146 +156,3 @@ def shell_exit_code(exit_code):
         return FAILURE
     else:
         return SUCCESS
-
-
-def declare_failed_by_kill(job, queue, sig):
-    """
-    Declare the job failed by a kill signal and put it in a suitable failed queue.
-    E.g. queue=queues.failed_data_in, if the kill signal was received during stage-in.
-
-    :param job: job object.
-    :param queue: queue object.
-    :param sig: signal.
-    :return:
-    """
-
-    job.state = 'failed'
-    error_code = errors.get_kill_signal_error_code(sig)
-    job.piloterrorcodes, job.piloterrordiags = errors.add_error_code(error_code)
-
-    queue.put(job)
-
-
-def scan_for_jobs(queues):
-    """
-    Scan queues until at least one queue has a job object. abort if it takes too long time
-
-    :param queues:
-    :return: found jobs (list of job objects).
-    """
-
-    t0 = time.time()
-    found_job = False
-    jobs = None
-
-    while time.time() - t0 < 30:
-        for q in queues._fields:
-            _q = getattr(queues, q)
-            jobs = list(_q.queue)
-            if len(jobs) > 0:
-                logger.info('found %d job(s) in queue %s after %d s - will begin queue monitoring' %
-                            (len(jobs), q, time.time() - t0))
-                found_job = True
-                break
-        if found_job:
-            break
-        else:
-            time.sleep(0.1)
-
-    return jobs
-
-
-def get_queuedata_from_job(queues):
-    """
-    Return the queuedata object from a job in the given queues object.
-    This function is useful if queuedata is needed from a function that does not know about the job object.
-    E.g. the pilot monitor does not know about the job object, but still knows
-    about the queues from which a job object can be extracted and therefore the queuedata.
-
-    :param queues: queues object.
-    :return: queuedata object.
-    """
-
-    queuedata = None
-
-    # extract jobs from the queues
-    jobs = scan_for_jobs(queues)
-    if jobs:
-        for job in jobs:
-            queuedata = job.infosys.queuedata
-            break
-
-    return queuedata
-
-
-def get_queuedata_from_job_old(queues):
-    """
-    Return the queuedata object from a job in the given queues object.
-    This function is useful if queuedata is needed from a function that does not know about the job object.
-    E.g. the pilot monitor does not know about the job object, but still knows
-    about the queues from which a job object can be extracted and therefore the queuedata.
-
-    :param queues: queues object.
-    :return: queuedata object.
-    """
-
-    # TODO: use scan_for_jobs() instead of the loop below
-
-    job = None
-    queuedata = None
-
-    # loop over all queues and find a job object, from which the queuedata object can be reached
-    delay = 5
-    maxretries = 5
-    for q in queues._fields:
-        _q = getattr(queues, q)
-        jobs = list(_q.queue)
-        logger.info('found %d job(s) in queue %s' % (len(jobs), q))
-        if len(jobs) == 0:
-            logger.info('taking a 5 s nap')
-            time.sleep(0.1)
-
-        retry = 0
-        for job in jobs:
-            # allow some time to pass in case no job object has been added to the queues yet
-            while retry < maxretries:
-                logger.info('attempting to extract queuedata from queues (#%d/#%d)' % (retry + 1, maxretries))
-                try:
-                    queuedata = job.infosys.queuedata
-                except Exception:
-                    logger.info('sleeping %d seconds to allow job object to be added to queues '
-                                '(needed for queuedata extraction)' % delay)
-                    time.sleep(delay)
-                    retry += 1
-                else:
-                    break
-        if queuedata:
-            break
-
-    return queuedata
-
-
-def abort_jobs_in_queues(queues, sig):
-    """
-    Find all jobs in the queues and abort them.
-
-    :param queues: queues object.
-    :param sig: detected kill signal.
-    :return:
-    """
-
-    jobs_list = []
-
-    # loop over all queues and find all jobs
-    for q in queues._fields:
-        _q = getattr(queues, q)
-        jobs = list(_q.queue)
-        for job in jobs:
-            if job not in jobs_list:
-                jobs_list.append(job)
-
-    logger.info('found %d job(s) in %d queues' % (len(jobs_list), len(queues._fields)))
-    for job in jobs_list:
-        log = get_logger(job.jobid)
-        log.info('aborting job %s' % (job.jobid))
-        declare_failed_by_kill(job, queues.failed_jobs, sig)
